pyproject.toml
[project]
name = "search-ranking-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.0",
    "python-dotenv>=1.2.1",
    "serpapi>=0.1.0",
    "uvicorn>=0.41.0",
]
